Suthesh/market-news-data (#2016)
* updated data source

* fixed key prop

* sort based on published date
